#11 - 프로필 컨트롤러와 회원 컨트롤러를 병합
프로필 컨트롤러를 따로 두니 `@RepositoryRestResource` 관련 오류가 발생함.
아마 프로필 도메인이 없어서 그런듯 하여 회원 컨트롤러에 합침
