+++ src/test/java/com/pinterest/controller/MemberControllerTest.java
@@ -10,13 +10,13 @@ import org.springframework.test.web.servlet.MockMvc;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
-@WebMvcTest(ProfileController.class)
-@DisplayName("View 컨트롤러 - 프로필")
-class ProfileControllerTest {
+@WebMvcTest(MemberController.class)
+@DisplayName("View 컨트롤러 - 회원")
+class MemberControllerTest {
 
     private final MockMvc mvc;
 
-    public ProfileControllerTest(@Autowired MockMvc mvc) {
+    public MemberControllerTest(@Autowired MockMvc mvc) {
         this.mvc = mvc;
     }
 
@@ -26,7 +26,7 @@ class ProfileControllerTest {
         // Given
 
         // When & Then
-        mvc.perform(get("/profile"))
+        mvc.perform(get("/members/1"))
                 .andExpect(status().isOk())
                 .andExpect(content().contentTypeCompatibleWith(MediaType.TEXT_HTML))
                 .andExpect(view().name("profile/index"))
